Changed boot time to unsigned long long.
        * glibtop_machine.h: Changed boot time to unsigned long long.

        * open.c: Added proper type checking for boot value. There's
        no point in saving a few nanoseconds in init function. And
        one day it will become 64-bit value.

        In glibtop_get_kstats(): Don't reread vminfo_snaptime when
        reinitializing kstats.

+++ sysdeps/solaris/glibtop_machine.h
@@ -51,13 +51,13 @@ struct _glibtop_machine
 
     kstat_t *cpu_stat_kstat [64];
 
-    kstat_t *system;	/* boot_time & avenrun* where needed */
-    kstat_t *syspages;	/* memory usage */
-    kstat_t *bunyip;	/* more memory usage */
+    kstat_t *system;		/* boot_time & avenrun* where needed */
+    kstat_t *syspages;		/* memory usage */
+    kstat_t *bunyip;		/* more memory usage */
 
-    int pagesize;	/* in kilobytes */
-    int ticks;		/* clock ticks, as returned by sysconf(_SC_CLK_TCK) */
-    unsigned boot;	/* boot time, it's ui32 in kstat */
+    int pagesize;		/* in kilobytes */
+    int ticks;			/* clock ticks, as returned by sysconf() */
+    unsigned long long boot;	/* boot time, it's ui32 in kstat */
 };
 
 END_LIBGTOP_DECLS
